requirements.txt
moviepy
PyPDF2
python-dotenv
pillow<10.0